requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
numpy
pillow
easyocr
datasets
rapidfuzz
scikit-learn